+++ threedi_sideview.py
@@ -47,7 +47,7 @@ class ThreeDiSideView:
         # initialize plugin directory
         self.plugin_dir = os.path.dirname(__file__)
 
-        self.icon_path = ':/plugins/ThreeDiToolbox/icon_route.png'
+        self.icon_path = ':/plugins/ThreeDiToolbox/icons/icon_route.png'
         self.menu_text = u'Show sideview of 3Di model with results'
 
         self.dock_widgets = []